real-time-processor/requirements.txt
kafka-python==2.0.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2
google-generativeai==0.3.1
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
cachetools==5.3.2
structlog>=24.1.0



